getResult now returns the actual token and no the raw response
